# Fix RecordKind populating