Fix #877

## Background

`Record.Kind` was not populated in `ReadRecordsFromDecompressedStream` execution path - so added it there

## Drawback

Missing test(s) on `ReadRecords`

This change will require the `ToBinaryLogRecordKind` to be kept properly in sync with possible new added EventArgs - otherwise `ReadRecords` might throw otherwise. So test would be nice to prevent about forgotting. (btw. this affects api only - not the GUI - it uses different reading methods)


WDYT @KirillOsenkov? Would you preffer this to already added fix (https://github.com/KirillOsenkov/MSBuildStructuredLog/commit/78d683efc99fa0c1b8eac14a5d21dfdbecc8ecfe)?