@KirillOsenkov 2025-07-03
This looks good, my only potential concern is the type switch might be more expensive than just passing the integer out.

But maybe my intuition is wrong and the cost will be completely drowned in disk access and decompression. 

Would you mind measuring ReadRecords for a large binlog with whatever is in main vs. your PR? Remember to warm up first, then measure. If there's no noticeable difference I'm happy to merge.