@JanKrivanek 2025-07-07
You were right @KirillOsenkov! (well - how could I doubt about your perf instinct ;-D)
There is a small - but measurable - perf hit by this PR method. So I'm abandoning this for now

It was fun experiment though :)

## Data

| LogName | Records # |
|------------|------------|
| msbuild-console.binlog cnt | 5928 |
| msbuild-10-projs.binlog cnt | 55952 |
| msbuild-50-projs.binlog cnt | 502567 |
| msbuild-100-projs.binlog cnt | 395290 |

## Main

| Method          | Mean       | Error     | StdDev     |
|---------------- |-----------:|----------:|-----------:|
| ReadConsoleLog  |   5.925 ms | 0.1182 ms |  0.1213 ms |
| Read10ProjsLog  |  56.885 ms | 1.1312 ms |  2.2329 ms |
| Read50ProjsLog  | 456.204 ms | 9.0176 ms | 18.4205 ms |
| Read100ProjsLog | 238.662 ms | 4.6952 ms |  7.1700 ms |

## This PR

| Method          | Mean       | Error     | StdDev    |
|---------------- |-----------:|----------:|----------:|
| ReadConsoleLog  |   6.221 ms | 0.0824 ms | 0.0771 ms |
| Read10ProjsLog  |  59.692 ms | 1.0683 ms | 1.8711 ms |
| Read50ProjsLog  | 469.211 ms | 4.9124 ms | 4.3547 ms |
| Read100ProjsLog | 252.555 ms | 4.9209 ms | 5.8580 ms |